+++ crates/merman-render/src/svg/parity.rs
@@ -36,6 +36,7 @@ mod radar;
 mod requirement;
 mod root_svg;
 mod roughjs46;
+mod roughjs_common;
 mod sankey;
 mod sequence;
 mod state;